requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
openai==1.68.2
psycopg2-binary==2.9.3
python-dotenv==0.19.0
gunicorn==20.1.0
bcrypt==3.2.0
werkzeug==2.0.3
SQLAlchemy==1.4.39
flask-babel==2.0.0
google-generativeai==0.7.2
PyPDF2==3.0.1
